// src/app/(app)/(tabs)/home/types/payment.types.ts
export interface PaymentInitPayload {
  userId: string | number;
  amount: number;
  investmentId: string | number;
  schemeId: string | number;
  userEmail: string;
  userMobile: string;
  userName: string;
  chitId: string | number;
}
// ...
export interface PaymentStatusUpdate {
  paymentResponse: {
    txn_id: string;
    amount: number;
    order_id: string;
    txn_detail: {
      status: 'CHARGED' | 'FAILED';
    };
    payment_gateway_response?: {
      resp_message: string;
    };
  };
}
// ...
export interface UserDetails {
  name?: string;
  accountname?: string;
  accNo?: string;
  accountNo?: string;
  mobile?: string;
  email?: string;
  userId?: string;
  investmentId?: string;
  schemeId?: string;
  chitId?: string;
  paymentFrequency?: string;
  schemeName?: string;
  isRetryAttempt?: boolean;
  retryData?: {
    paymentData: {
      userId: string;
      investmentId: string;
      schemeId: string;
      chitId: string;
      amount: number;
    };
    investmentData: {
      accountName: string;
      accountNo: string;
    };
  };
  userEmail?: string;
  userMobile?: string;
  data?: {
    data?: {
      userId?: string;
      id?: string;
      schemeId?: string;
      chitId?: string;
    };
  };
}